moved differentiation rules into a separate file. All tests pass except.
Need to replace ifelse with my own ifelse that generates correct if...else code. Too many functions execute illegal stuff in one of the ifelse branches which causes an exception.
